Adds usage examples, changes file-not-found behaviour
This adds usage examples in the form of the exam.mk makefile (fixes #26).
This also adds some example of alternative use (fixes #24).

We also change the error handling if a database file does not exist. We
will not terminate if a file does not exist, we now output an error and
continue. This will allow us to use upcoming exams as databases, so we
only have to update the Makefile once a year.