Redirects std{in,out,err} to TTY for editor, prints to stderr
The editor might require that std{in,out,err} are connected to the
controlling terminal (e.g. Vim). This ensures that happens.

We also want to redirect stdout to a file to store the questions, thus
we only want to print the generated questions to stdout. This changes
all other output to stderr.